ENGR00233452 Integrate 4.6.9p9 GPU kernel part driver
Acked-by: Lily Zhang

+++ drivers/mxc/gpu-viv/hal/os/linux/kernel/gc_hal_kernel_debugfs.c
@@ -167,7 +167,7 @@ static gcsDebugFileSystem gc_dbgfs ;
 static caddr_t
 _ReadFromNode (
                 gcsDebugFileSystemNode* Node ,
-                int *Length ,
+                size_t *Length ,
                 loff_t *Offset
                 )
 {
